% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCLUST.R
\name{HCLUST}
\alias{HCLUST}
\title{Perform basic heirarchical clustering on a data matrix using single linkage.}
\usage{
HCLUST(data.matrix, K)
}
\arguments{
\item{data.matrix}{A numeric data matrix to perform clustering on}

\item{K}{The number of clusters}
}
\value{
A vector with as many elements as data.matrix has rows where each
element indicate which cluster the row at that index in the
data.matrix is assigned, with there being K different clusters
represented in the vector
}
\description{
This function computes its own distance matrix.
}
