% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMEANS.R
\name{KMEANS}
\alias{KMEANS}
\title{Perform basic k-means clustering on a data matrix}
\usage{
KMEANS(data.matrix, K)
}
\arguments{
\item{data.matrix}{A numeric data matrix to perform clustering on}

\item{K}{The number of clusters}
}
\value{
A list which has the following elements
cluster: a vector of integers (from 1 to k) indicating the cluster
to which each row of the data matrix is associated
centers: a matrix of K cluster centers
withinss: a vector of the within-cluster sum squares for each
cluster
tot.withinss: the sum of withinss for all clusters
size: a vector of integers indicating the number of points
associated with each cluster
iter: the number of outer iterations the function took to converge
}
\description{
Perform basic k-means clustering on a data matrix
}
